requirements.txt
firebase-admin
google-generativeai
sendgrid
python-dotenv